feat: rework package ids and refs
Previously, package ids were
(PackageNames, VersionId). This worked, but it was
prone to source collisions. Additionally, it
simply doesn't work with decentralised pesde. This
commit reworks package ids to be
(PackageSources, PackageRefs, VersionId). This
enforces scoping of packages to their sources
while not relying on package names (which are
not present in decentralised pesde).

As for package refs, they have been reworked to
store the source-specific information needed to
locate a package in a given source. For pesde and
Wally, this is simply the package name. For Git,
this is the revision and type of package
(pesde, Wally). This also eliminates redundancy
between refs and sources - previously, the source
AND the ref contained the repository URL for
Git-based sources. This led to internal
inconsistencies and made reasoning about sources
difficult. Now, since the source is stored in the
id, the ref only needs to store what neither the
source nor the version id store.