feat: force git URLs to use https scheme
This change allows, and even strongly encourages,
git URLs to be specified without a scheme. Unless
the scheme is "https://" it will be rejected to
simplify migration. In pesde 0.9 all schemes will
be rejected.

This will help with caching and avoid issues with
different schemes causing multiple clones of the
package. Additionally, it simplifies overriding
the URL scheme using Git configuration, as now
required by pesde in order to use other protocols.